custom_vision/requirements.txt
azure-cognitiveservices-vision-customvision
annotationfactory